best practice

+++ src/main/java/com/build4all/booking/repository/BookingRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
@@ -36,7 +37,7 @@ public interface BookingRepository extends JpaRepository<Booking, Long> {
            JOIN ib.item i
            WHERE i.business.id = :businessId
            """)
-    long countByBusinessId(@Param("businessId") Long businessId);
+    BigDecimal countByBusinessId(@Param("businessId") Long businessId);
 
     @Query("""
            SELECT DISTINCT b
